created like model and working on like view

+++ app/controllers/likes_controller.rb
@@ -3,6 +3,7 @@ class LikesController < ApplicationController
   def create
     @bookmark = Bookmark.find(params[:bookmark_id])
     @like = current_user.likes.build(bookmark: @bookmark)
+    authorize @like
 
     if @like.save
       flash[:notice] = "Bookmark liked."
@@ -16,6 +17,8 @@ class LikesController < ApplicationController
   def destroy
     @bookmark = Bookmark.find(params[:bookmark_id])
     @like = @bookmark.likes.find(params[:id])
+    authorize @like
+    
     if @like.destroy
       flash[:notice] = "Like deleted."
       redirect_to [@bookmark.topic, @bookmark]
